pyproject.toml
[project]
name = "advancedif-skill-benchmark"
version = "0.1.0"
description = "AdvancedIF answer-refinement benchmark with restricted judge feedback, iterative and RLM harnesses, and carried skill artifacts."
readme = "README.md"
requires-python = ">=3.11"
tags = [
    "advancedif",
    "verifiers",
    "prime-rl",
    "rlm",
    "partial-feedback",
    "skill",
    "judge",
    "instruction-following",
]
dependencies = [
    "datasets>=4.0.0",
    "openai>=1.108.1",
    "python-dotenv>=1.0.0",
    "verifiers>=0.1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = [
    "advancedif_iter_skill.py",
    "advancedif_rlm_skill.py",
    "sequence_runner.py",
    "run_initial_study.py",
    "run_study.py",
    "run_research_suite.py",
    "core",
]

[tool.uv.sources]
verifiers = { path = "../verifiers" }

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = [
    "advancedif_iter_skill.py",
    "advancedif_rlm_skill.py",
    "sequence_runner.py",
    "run_initial_study.py",
    "run_study.py",
    "run_research_suite.py",
    "core",
    "tests",
]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
